feat(typography): prop types and apply styling
