+++ packages/typography/src/index.tsx
@@ -1,0 +1,20 @@
+import { createElement } from 'react';
+import { useStyles } from './style';
+import { TypographyType } from './types';
+
+export type TypographyProps<K extends keyof T, T extends TypographyType = TypographyType> = {
+  variant: K;
+  type: keyof T[K];
+  component?: keyof TypographyType;
+};
+
+export const Typography = <K extends keyof T, T extends TypographyType>(
+  props: React.PropsWithChildren<TypographyProps<K, T>>
+) => {
+  const { variant, type, component, children } = props;
+  const styles = useStyles(variant, type)();
+  const el = createElement(component || String(variant as keyof T), { className: styles.root }, children);
+  return el;
+};
+
+export default Typography;
