+++ packages/typography/src/style.ts
@@ -1,14 +1,20 @@
-import { createStyles, makeStyles, FusionTheme } from '@equinor/fusion-react-styles';
+import { createStyles, makeStyles } from '@equinor/fusion-react-styles';
+import { TypographyType } from './types';
 
-export const useStyles = <T extends FusionTheme['typography'], K extends keyof T>(variant: K, type: T[K]) => {
+export const useStyles = <T extends TypographyType, K extends keyof T>(variant: keyof T, type: keyof T[K]) => {
   return makeStyles(
     (theme) =>
       createStyles({
-        // type: (props: { variant: keyof FusionTheme['typography']; type: TypographyType }) => ({
-        type: (props) => ({
-          ...theme.typography[variant][type].styles,
-          //   ...theme.typography['heading']['h1'].style,
-        }),
+        root: () => {
+          const typography = theme.typography[variant as keyof typeof theme.typography];
+          const styleProperty = typography[type as keyof typeof typography] || { style: {} };
+          return {
+            // eslint-disable-next-line @typescript-eslint/ban-ts-comment
+            // @ts-ignore
+            ...styleProperty.style,
+            color: 'red',
+          };
+        },
       }),
     { name: 'fusion-typography' }
   );
