+++ packages/typography/src/types.ts
@@ -1,0 +1,4 @@
+import { FusionTheme } from '@equinor/fusion-react-styles';
+
+export type TypographyType = FusionTheme['typography'];
+export type TypographyHeadingType = TypographyType['heading'];
